requirements.txt
flask==2.2.3
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-redis==0.39b0
opentelemetry-instrumentation-requests==0.39b0
redis==4.5.4
requests==2.28.2
